📝 updated subscriptions method save file
